+++ app/renderer/epics/homePage.epics.js
@@ -1,15 +1,19 @@
 import { EMPTY, from, of } from 'rxjs';
-import { mergeMap } from 'rxjs/operators';
+import { mergeMap, catchError, map } from 'rxjs/operators';
 import { ofType } from 'redux-observable';
 import types from 'renderer/types';
 import ROUTES from 'renderer/constants/routes';
 import { storageKeys, getStorage } from 'renderer/api/utils/storage';
 import { NetworkNameOptions } from 'common/types/cennznet-node.types';
+import { Logger } from 'renderer/utils/logging';
+import { Observable } from 'rxjs/Observable';
+import { cennznetStatusChannel } from 'renderer/ipc/cennznet.ipc';
 
-const homePageLoadEpic = (action$, state$) =>
+const homePageNavigationEpic = (action$, state$) =>
   action$.pipe(
-    ofType(types.homePageLoad.triggered),
+    ofType(types.homePageNavigation.triggered),
     mergeMap(async () => {
+      Logger.debug(`HomePageNavigationEpic call::`);
       const isTosAccepted = await getStorage(storageKeys.TERMS_OF_USE_ACCEPTANCE);
       const isNetworkRemembered = await getStorage(storageKeys.REMEMBER_NETWORK);
       const selectedNetwork = await getStorage(storageKeys.SELECTED_NETWORK);
@@ -20,16 +24,20 @@ const homePageLoadEpic = (action$, state$) =>
         return { type: types.navigation.triggered, payload: ROUTES.TERMS_OF_USE_ACCEPTANCE };
       }
 
-      if (isNetworkRemembered && selectedNetwork) {
-        if (
-          (selectedNetwork === NetworkNameOptions.LOCAL_TESTNET && genesisConfigFilePath) ||
-          selectedNetwork !== NetworkNameOptions.LOCAL_TESTNET
-        )
-          return { type: types.navigation.triggered, payload: ROUTES.SYNC_NODE };
+      if (
+        isNetworkRemembered &&
+        ((selectedNetwork === NetworkNameOptions.LOCAL_TESTNET && genesisConfigFilePath) ||
+          selectedNetwork !== NetworkNameOptions.LOCAL_TESTNET)
+      ) {
+        return { type: types.navigation.triggered, payload: ROUTES.SYNC_NODE };
       }
 
       return { type: types.navigation.triggered, payload: ROUTES.CHOOSE_NETWORK };
+    }),
+    catchError(err => {
+      Logger.debug(`Error in HomePageNavigationEpic: ${err}`);
+      return of({ type: types.navigation.triggered, payload: ROUTES.TERMS_OF_USE_ACCEPTANCE });
     })
   );
 
-export default [homePageLoadEpic];
+export default [homePageNavigationEpic];