+++ app/renderer/pages/homePage/index.js
@@ -1,8 +1,9 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import Button from 'components/Button';
 import styled from 'styled-components';
 import Spinner from 'components/Spinner';
 import { Layout, LayoutWrapper, MainContent } from 'components/layout';
+import { cennznetStatusChannel } from 'renderer/ipc/cennznet.ipc';
 import SideNav from 'components/layout/SideNav';
 import SimpleSidebar from 'components/layout/SimpleSidebar';
 import { environment } from 'common/environment';
@@ -17,7 +18,13 @@ const SpinnerWrapper = styled.div`
   display: flex;
 `;
 
-const HomePage = ({ hasBlockNumbers }) => {
+const HomePage = ({ hasBlockNumbers, onPageNavigation, onSubscribeCennznetStatus }) => {
+  useEffect(() => {
+    if (hasBlockNumbers) {
+      onPageNavigation();
+    }
+  }, [hasBlockNumbers]);
+
   return (
     <Layout sidebar={isDevOrDebugProd ? <SideNav /> : <SimpleSidebar />}>
       <LayoutWrapper>