+++ CHANGELOG.md
@@ -8,11 +8,13 @@
 ### Fixes
 
 ### Improvements
+
 - ready for github release
 
 # v0.6.1 (2019.03.12)
 
 ### Fixes
+
 - add local net option back
 
 # v0.6.0 (2019.03.12)
@@ -35,6 +37,7 @@
 
 ### Fixes
 
+- loading issue in homePage
 - fix setInterval timer issue
 
 ### Improvements